Rewrite most of the things.
This time I take a great simplification of everything and getting rid of
__index and __newindex metamethods. Instead I use an __index table.

As a downside, it means it's no longer possible to make syntax sugar for
fields/properties (foo.x = 10). But it's not bad at all, in fact it's a good
thing, because this syntax sugar hides potentially costly operation (if 'x' is
a userdata, it'll be malloced each time).

Using __index table has good sides as well. The best part is that it is very
fast. VM doesn't have to put its values onto the stack in order to negotiate
things with C code anymore. And the code follows KISS very nicely.